backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.0
openai>=1.0.0
pillow>=12.0.0
PyMuPDF>=1.24.0
python-pptx>=1.0.0
python-dotenv>=1.0.1
img2pdf>=0.5.1
reportlab>=4.1.0
tenacity>=9.0.0
python-docx>=1.1.0
werkzeug>=3.0.1
